Stop relying on the @resin.io/device-types module
Change-type: patch
