+++ src/features/device-types/device-type-json.ts
@@ -1,0 +1,64 @@
+export interface DeviceTypeJson {
+	slug: string;
+	name: string;
+	aliases?: string[];
+	logoUrl?: string;
+
+	arch: string;
+	state?: string;
+
+	isDependent?: boolean;
+	instructions?: string[] | DeviceTypeInstructions;
+	gettingStartedLink?: string | DeviceTypeGettingStartedLink;
+	stateInstructions?: { [key: string]: string[] };
+	options?: Array<DeviceTypeOptions | DeviceTypeOptionsGroup>;
+	initialization?: {
+		options?: DeviceInitializationOptions[];
+		operations: Array<{
+			command: string;
+		}>;
+	};
+	supportsBlink?: boolean;
+	yocto: {
+		fstype?: string;
+		deployArtifact: string;
+	};
+	// Holds the latest resinOS version
+	buildId: string;
+}
+interface DeviceTypeInstructions {
+	linux: string[];
+	osx: string[];
+	windows: string[];
+}
+
+interface DeviceTypeGettingStartedLink {
+	linux: string;
+	osx: string;
+	windows: string;
+	[key: string]: string;
+}
+
+interface DeviceTypeOptionBase {
+	message: string;
+	name: string;
+}
+
+interface DeviceTypeOptions extends DeviceTypeOptionBase {
+	options: DeviceTypeOptionsGroup[];
+	collapsed?: boolean;
+	isCollapsible?: boolean;
+	isGroup: boolean;
+}
+
+interface DeviceInitializationOptions extends DeviceTypeOptionBase {
+	type: string;
+}
+
+interface DeviceTypeOptionsGroup extends DeviceTypeOptionBase {
+	default?: number | string;
+	type: string;
+	min?: number;
+	choices?: string[] | number[];
+	choicesLabels?: { [key: string]: string };
+}
