+++ src/index.ts
@@ -151,7 +151,7 @@ export * as tags from './features/tags/validation';
 export type { Creds, User } from './infra/auth/jwt-passport';
 export type { Access } from './features/registry/registry';
 export type { ApplicationType } from './features/application-types/application-types';
-export type { DeviceTypeJson } from './features/device-types/device-types';
+export type { DeviceTypeJson } from './features/device-types/device-type-json';
 
 export { DefaultApplicationType } from './features/application-types/application-types';
 export * as request from './infra/request-promise';
